+++ src/main/java/com/marioborrego/api/calculodeduccionesbackend/personal/presentation/dto/bbcc/ActualizarBbccPersonalDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class ActualizarBbccPersonalDTO {
-    private int idBbccPersonal;
+    private Long idBbccPersonal;
     private String campoActualizado;
     private Double valor;
 }
